moved the musics class into a seperate file
